TalkFamous/src/analytics_insights.py
```python
import pandas as pd
from sklearn import metrics
from utils import load_data

# Load shared data
user_data = load_data("user_data")
creator_data = load_data("creator_data")
analytics_data = load_data("analytics_data")

class AnalyticsInsights:
    def __init__(self, user_data, creator_data, analytics_data):
        self.user_data = user_data
        self.creator_data = creator_data
        self.analytics_data = analytics_data

    def track_user_engagement(self):
        # Track user engagement based on interactions with the app
        user_engagement = self.user_data.groupby('user_id')['interaction_count'].sum()
        return user_engagement

    def track_app_usage(self):
        # Track app usage based on login frequency and duration
        app_usage = self.user_data.groupby('user_id')['login_frequency', 'login_duration'].mean()
        return app_usage

    def analyze_audience_behavior(self):
        # Analyze audience behavior based on their interactions and preferences
        audience_behavior = self.user_data.groupby('user_id')['interaction_type', 'preference'].agg(pd.Series.mode)
        return audience_behavior

    def generate_insights(self):
        # Generate insights based on user engagement, app usage, and audience behavior
        user_engagement = self.track_user_engagement()
        app_usage = self.track_app_usage()
        audience_behavior = self.analyze_audience_behavior()

        insights = pd.concat([user_engagement, app_usage, audience_behavior], axis=1)
        return insights

    def provide_recommendations(self):
        # Provide recommendations based on the generated insights
        insights = self.generate_insights()
        recommendations = insights.apply(lambda row: self.recommend(row), axis=1)
        return recommendations

    def recommend(self, row):
        # Recommend strategies based on the insights
        if row['interaction_count'] < 5:
            return 'Increase engagement'
        elif row['login_frequency'] < 3:
            return 'Increase app usage'
        elif row['preference'] != 'AI voice memes':
            return 'Promote AI voice memes'
        else:
            return 'Maintain current strategy'

analytics_insights = AnalyticsInsights(user_data, creator_data, analytics_data)
recommendations = analytics_insights.provide_recommendations()
print(recommendations)
```